Same as above only in this case it wont try to delete

+++ api/middlewares/clearUploadsFolder.js
@@ -3,13 +3,17 @@ const path = require('path');
 
 const directory = path.join(__dirname, '../../uploads');
 module.exports = (req, res, next)=>{
-					fs.readdir(directory, (err, files)=>{
-					   	if(err) throw err;
-				        	for (const file of files) {
-				        		fs.unlink(path.join(directory, file), err=>{
-				        			if(err) throw err;
-				        		});
-				        	}
-				        });
+					if (process.env.NODE_ENV === 'production') {
 				        next();
+					}else {
+						fs.readdir(directory, (err, files)=>{
+						   	if(err) throw err;
+					        	for (const file of files) {
+					        		fs.unlink(path.join(directory, file), err=>{
+					        			if(err) throw err;
+					        		});
+					        	}
+					        });
+				        next();
+					}
 				};
